requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.2.1
colorama==0.4.6
distro==1.9.0
Django==5.1.2
django-cors-headers==3.14.0
django-model-utils==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
fastapi==0.116.1
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-core==0.3.74
langchain-text-splitters==0.3.9
langsmith==0.4.14
openai==1.99.9
orjson==3.11.2
packaging==25.0
Pillow==10.1.0
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.0
python-decouple==3.8
python-dotenv==1.1.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.43
sqlparse==0.5.2
starlette==0.47.2
tenacity==9.1.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.35.0
zstandard==0.23.0
